Code refactoring.
Move AATAFilePipeline into pipeline.projects.aata.
Move make_la_person into pipeline.linkedart.
